@@ -11,40 +11,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.1.3",
-    "@angular/cdk": "8.1.2",
-    "@angular/common": "8.1.3",
-    "@angular/compiler": "8.1.3",
-    "@angular/core": "8.1.3",
-    "@angular/forms": "8.1.3",
-    "@angular/material": "8.1.2",
-    "@angular/platform-browser": "8.1.3",
-    "@angular/platform-browser-dynamic": "8.1.3",
-    "@angular/router": "8.1.3",
-    "rxjs": "6.4.0",
-    "tslib": "1.10.0",
+    "@angular/animations": "9.1.0",
+    "@angular/cdk": "9.2.0",
+    "@angular/common": "9.1.0",
+    "@angular/compiler": "9.1.0",
+    "@angular/core": "9.1.0",
+    "@angular/forms": "9.1.0",
+    "@angular/material": "9.2.0",
+    "@angular/platform-browser": "9.1.0",
+    "@angular/platform-browser-dynamic": "9.1.0",
+    "@angular/router": "9.1.0",
+    "highcharts": "^8.0.4",
+    "jwt-decode": "^2.2.0",
+    "rxjs": "~6.5.4",
+    "tslib": "1.11.1",
     "uuid": "3.3.2",
-    "zone.js": "0.9.1"
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.801.3",
-    "@angular/cli": "8.1.3",
-    "@angular/compiler-cli": "8.1.3",
-    "@angular/language-service": "8.1.3",
-    "@types/node": "8.9.5",
+    "@angular-devkit/build-angular": "^0.901.0",
+    "@angular/cli": "9.1.0",
+    "@angular/compiler-cli": "9.1.0",
+    "@angular/language-service": "9.1.0",
     "@types/jasmine": "3.3.16",
     "@types/jasminewd2": "2.0.6",
+    "@types/node": "^12.11.1",
     "codelyzer": "5.1.0",
     "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.1.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "2.2.0",
     "karma-coverage-istanbul-reporter": "2.0.6",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.2",
     "protractor": "5.4.2",
-    "ts-node": "7.0.1",
-    "tslint": "5.15.0",
-    "typescript": "3.4.5"
+    "ts-node": "8.3.0",
+    "tslint": "6.1.0",
+    "typescript": "~3.8.3"
   }
 }
