@@ -1,25 +1,23 @@
-import {Component} from '@angular/core';
-import {AbstractControl, FormControl, FormGroup, Validators} from '@angular/forms';
-import {GameService} from '../../../services/game.service';
-import {GameFormGroup, MAX_GOALS, MIN_GOALS} from '../../../constants';
-import {teamValidator, playerGoalsValidator, goalsValidator, playerValidator} from '../../../validators/game-validator';
-import {Editor} from '../editor';
-import {IGame} from '../../../types';
-import {MatSnackBar} from '@angular/material';
+import { Component } from '@angular/core';
+import { AbstractControl, FormControl, FormGroup, Validators } from '@angular/forms';
+import { GameService } from '../../../services/game.service';
+import { GameFormGroup, MAX_GOALS, MIN_GOALS } from '../../../constants';
+import { teamValidator, playerGoalsValidator, goalsValidator, playerValidator } from '../../../validators/game-validator';
+import { Editor } from '../editor';
+import { IGame } from '../../../types';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
 
 @Component({
-    selector: 'app-editor-game',
-    templateUrl: './editor-game.component.html',
-    styleUrls: ['./editor-game.component.css']
+    selector: 'app-editor-game', templateUrl: './editor-game.component.html', styleUrls: [ './editor-game.component.css' ]
 })
 export class EditorGameComponent extends Editor<IGame> {
     readonly maxGoals = MAX_GOALS;
     constructor(private gameService: GameService, snackBar: MatSnackBar) {
         super(gameService, snackBar);
         this.formGroup = new FormGroup({
-            team1: this.prepareTeamGroup(),
-            team2: this.prepareTeamGroup()
-        }, {validators: [teamValidator, playerGoalsValidator, goalsValidator, playerValidator]});
+            team1: this.prepareTeamGroup(), team2: this.prepareTeamGroup()
+        }, { validators: [ teamValidator, playerGoalsValidator, goalsValidator, playerValidator ] });
     }
 
     get team1(): AbstractControl {
@@ -47,13 +45,13 @@ export class EditorGameComponent extends Editor<IGame> {
     }
 
     isNumber(event: KeyboardEvent): boolean {
-        return event.charCode >= '0'.charCodeAt(0)  && event.charCode <= '9'.charCodeAt(0);
+        return event.charCode >= '0'.charCodeAt(0) && event.charCode <= '9'.charCodeAt(0);
     }
 
     private prepareTeamGroup(): FormGroup {
         return new FormGroup({
-            id: new FormControl(undefined, [Validators.required]),
-            goals: new FormControl(MIN_GOALS, [Validators.required, Validators.max(MAX_GOALS), Validators.min(MIN_GOALS)]),
+            id: new FormControl(undefined, [ Validators.required ]),
+            goals: new FormControl(MIN_GOALS, [ Validators.required, Validators.max(MAX_GOALS), Validators.min(MIN_GOALS) ]),
             player1: this.preparePlayerGroup(),
             player2: this.preparePlayerGroup()
         });
@@ -61,8 +59,8 @@ export class EditorGameComponent extends Editor<IGame> {
 
     private preparePlayerGroup(): FormGroup {
         return new FormGroup({
-            id: new FormControl(undefined, [Validators.required]),
-            goals: new FormControl(MIN_GOALS, [Validators.max(MAX_GOALS), Validators.min(MIN_GOALS)])
+            id: new FormControl(undefined, [ Validators.required ]),
+            goals: new FormControl(MIN_GOALS, [ Validators.max(MAX_GOALS), Validators.min(MIN_GOALS) ])
         });
     }
 }
