@@ -1,14 +1,27 @@
-import {Component, EventEmitter, Output} from '@angular/core';
+import { Component, EventEmitter, Output } from '@angular/core';
+import { AuthService } from '../../services/auth.service';
+import { Router } from '@angular/router';
+
 
 @Component({
-    selector: 'app-navigation',
-    templateUrl: './navigation.component.html',
-    styleUrls: ['./navigation.component.css']
+    selector: 'app-navigation', templateUrl: './navigation.component.html', styleUrls: [ './navigation.component.css' ]
 })
 export class NavigationComponent {
     @Output() selectLink: EventEmitter<void> = new EventEmitter();
 
-    closeSidenav(): void {
+    constructor(private authService: AuthService, private router: Router) {}
+
+    get isEnabled() {
+        return !!this.authService.token;
+    }
+
+    closeSidenav() {
         this.selectLink.emit();
+    }
+
+    async exit() {
+        this.closeSidenav();
+        this.authService.removeToken();
+        await this.router.navigateByUrl('login')
     }
 }
