@@ -1,15 +1,17 @@
-import {Component} from '@angular/core';
-import {PlayerService} from '../../services/player.service';
-import {TeamService} from '../../services/team.service';
-import {ITeam, ITeamUser, IUser} from '../../types';
-import {MatDialog, MatSnackBar} from '@angular/material';
-import {DialogComponent} from '../dialog/dialog.component';
-import {Message, SNACK_BAR_DURATION} from '../../constants';
+import { Component } from '@angular/core';
+import { PlayerService } from '../../services/player.service';
+import { TeamService } from '../../services/team.service';
+import { ITeam, ITeamUser, IUser } from '../../types';
+import { DialogComponent } from '../dialog/dialog.component';
+import { Message, SNACK_BAR_DURATION } from '../../constants';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
 
 @Component({
     selector: 'app-team-player-drag-and-drop',
     templateUrl: './team-player-drag-and-drop.component.html',
-    styleUrls: ['./team-player-drag-and-drop.component.css']
+    styleUrls: [ './team-player-drag-and-drop.component.css' ]
 })
 export class TeamPlayerDragAndDropComponent {
     teams: ITeam[] = [];
@@ -20,13 +22,16 @@ export class TeamPlayerDragAndDropComponent {
     constructor(private playerService: PlayerService,
                 private teamService: TeamService,
                 public dialog: MatDialog,
-                private snackBar: MatSnackBar) {
-        this.playerService.getAll().subscribe(players => {
-            this.players = players;
-        });
-        this.teamService.getAll().subscribe(teams => {
-            this.teams = teams;
-        });
+                private snackBar: MatSnackBar
+    ) {
+        this.playerService.getAll()
+            .subscribe(players => {
+                this.players = players;
+            });
+        this.teamService.getAll()
+            .subscribe(teams => {
+                this.teams = teams;
+            });
 
         this.accept = this.accept.bind(this);
         this.onFailed = this.onFailed.bind(this);
@@ -57,12 +62,12 @@ export class TeamPlayerDragAndDropComponent {
     openDialog(team: ITeam): void {
         const dialogRef = this.dialog.open(DialogComponent, {
             data: {
-                user: this.selectedPlayer,
-                team
+                user: this.selectedPlayer, team
             }
         });
 
-        dialogRef.afterClosed().subscribe(this.accept);
+        dialogRef.afterClosed()
+            .subscribe(this.accept);
     }
 
     private accept(result: ITeamUser): void {
