@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {IUser} from '../types';
+import { ITeam, IUser } from '../types';
 import {UrlAddress} from '../constants';
 import {HttpClient} from '@angular/common/http';
 import {EditorService} from './editor';
@@ -10,11 +10,15 @@ export class PlayerService extends EditorService<IUser> {
         super(httpClient, UrlAddress.user);
     }
 
-    updateTeams(id: number, team): void {
+    updateTeams(id: string, team: ITeam): void {
         this.data.value.find(player => player.id === id).teams.push({
             id: team.id,
             name: team.name,
             image: team.image
         });
+    }
+
+    updateUser(id: string, data) {
+        return this.httpClient.put(`${this.urlAddress}/${id}`, data);
     }
 }
